+++ experimental/packages/opentelemetry-sdk-node/src/index.ts
@@ -25,5 +25,6 @@ export * as metrics from '@opentelemetry/sdk-metrics';
 export * as node from '@opentelemetry/sdk-trace-node';
 export * as resources from '@opentelemetry/resources';
 export * as tracing from '@opentelemetry/sdk-trace-base';
-export { LoggerProviderConfig, MeterProviderConfig, NodeSDK } from './sdk';
-export { NodeSDKConfiguration } from './types';
+export { NodeSDK } from './sdk';
+export type { LoggerProviderConfig, MeterProviderConfig } from './sdk';
+export type { NodeSDKConfiguration } from './types';